+++ src/main/java/io/xlorey/fluxloader/server/core/Core.java
@@ -1,7 +1,10 @@
 package io.xlorey.fluxloader.server.core;
 
+import io.xlorey.fluxloader.enums.EventPriority;
 import io.xlorey.fluxloader.plugin.PluginManager;
 import io.xlorey.fluxloader.plugin.PluginRegistry;
+import io.xlorey.fluxloader.server.handlers.OnServerShutdownHandler;
+import io.xlorey.fluxloader.shared.EventManager;
 import io.xlorey.fluxloader.utils.Logger;
 
 /**
@@ -39,6 +42,8 @@ public class Core {
         Logger.print("FluxLoader Core initialization for the server..");
 
         PluginManager.loadPlugins(false);
+
+        EventManager.subscribe(new OnServerShutdownHandler(), EventPriority.HIGHEST);
 
         PluginManager.executePlugins(PluginRegistry.getLoadedServerPlugins());
     }